requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=1.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
httpx>=0.24.0
redis>=4.0.0
celery>=5.2.0
structlog>=21.1.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.0
pinecone-client>=2.2.0
numpy>=1.24.0
flower>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0
prometheus-client>=0.17.0 